+++ src/execution/utils.ts
@@ -92,6 +92,7 @@ import type {
   RhinestoneConfig,
   SignerSet,
   SourceAssetInput,
+  Sponsorship,
   TokenRequest,
   TokenSymbol,
   Transaction,
@@ -173,7 +174,7 @@ async function prepareTransaction(
     transaction.gasLimit,
     tokenRequests,
     recipient,
-    sponsored ?? false,
+    sponsored,
     eip7702InitSignature,
     settlementLayers,
     sourceAssets,
@@ -638,7 +639,7 @@ async function prepareTransactionAsIntent(
   gasLimit: bigint | undefined,
   tokenRequests: TokenRequest[],
   recipientInput: RhinestoneAccountConfig | Address | undefined,
-  isSponsored: boolean,
+  sponsored: Sponsorship | undefined,
   eip7702InitSignature: Hex | undefined,
   settlementLayers: SettlementLayer[] | undefined,
   sourceAssets: SourceAssetInput | undefined,
@@ -700,11 +701,19 @@ async function prepareTransactionAsIntent(
     options: {
       topupCompact: lockFunds ?? false,
       feeToken: feeAsset,
-      sponsorSettings: {
-        gasSponsored: isSponsored,
-        bridgeFeesSponsored: isSponsored,
-        swapFeesSponsored: isSponsored,
-      },
+      sponsorSettings: sponsored
+        ? typeof sponsored === 'object'
+          ? {
+              gasSponsored: sponsored.gas,
+              bridgeFeesSponsored: sponsored.bridging,
+              swapFeesSponsored: sponsored.swaps,
+            }
+          : {
+              gasSponsored: sponsored,
+              bridgeFeesSponsored: sponsored,
+              swapFeesSponsored: sponsored,
+            }
+        : undefined,
       settlementLayers,
     },
   }