+++ src/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @rhinestone/sdk
 
+## 1.0.32
+
+### Patch Changes
+
+- d42a5ac: Sponsorship types
+
+## 1.0.31
+
+### Patch Changes
+
+- 53c915e: Fix error in `toViewOnlyAccount` util
+
+## 1.0.30
+
+### Patch Changes
+
+- 5e7d4f5: Account deployment action
+
 ## 1.0.29
 
 ### Patch Changes